requirements.txt
geopandas 
shapely 
osmnx
folium
geopandas